requirements.txt
ollama
structlog
clickhouse-connect
numpy<2.0
pandas
sqlglot
sqlparse
presidio-analyzer
presidio-anonymizer
python-dotenv
tabulate
pyyaml
jsonschema
sqlalchemy
jinja2
flask
flask-cors
redis>=5.0.0
requests>=2.31.0
pytest
pytest-cov
pytest-mock
